Add option --keep-physical to remove backups just from nixops state, and leave the physical copy as-is. When using --force with 'nixops backup' do not perform status checks on existing backups.
(cherry picked from commit 11a127cf953fdabd92b83a14aebf1ca59268977c)

Conflicts:
	scripts/nixops

+++ nixops/backends/ec2.py
@@ -349,22 +349,23 @@ class EC2State(MachineState):
         return backups
 
 
-    def remove_backup(self, backup_id):
+    def remove_backup(self, backup_id, keep_physical = False):
         self.log('removing backup {0}'.format(backup_id))
         self.connect()
         _backups = self.backups
         if not backup_id in _backups.keys():
             self.warn('backup {0} not found, skipping'.format(backup_id))
         else:
-            for dev, snapshot_id in _backups[backup_id].items():
-                snapshot = None
-                try:
-                    snapshot = self._get_snapshot_by_id(snapshot_id)
-                except:
-                    self.warn('snapshot {0} not found, skipping'.format(snapshot_id))
-                if not snapshot is None:
-                    self.log('removing snapshot {0}'.format(snapshot_id))
-                    nixops.ec2_utils.retry(lambda: snapshot.delete())
+            if not keep_physical:
+                for dev, snapshot_id in _backups[backup_id].items():
+                    snapshot = None
+                    try:
+                        snapshot = self._get_snapshot_by_id(snapshot_id)
+                    except:
+                        self.warn('snapshot {0} not found, skipping'.format(snapshot_id))
+                    if not snapshot is None:
+                        self.log('removing snapshot {0}'.format(snapshot_id))
+                        nixops.ec2_utils.retry(lambda: snapshot.delete())
 
             _backups.pop(backup_id)
             self.backups = _backups
